Add module to handle book functionalities
