+++ .rubocop.yml
@@ -1,9 +1,9 @@
 AllCops:
   NewCops: enable
   Exclude:
-    - "Guardfile"
-    - "Rakefile"
-    - "node_modules/**/*"
+    - 'Guardfile'
+    - 'Rakefile'
+    - 'node_modules/**/*'
 
   DisplayCopNames: true
 
@@ -16,9 +16,8 @@ Metrics/AbcSize:
 Metrics/ClassLength:
   Max: 250
 Metrics/BlockLength:
-  IgnoredMethods: ['describe']
+  AllowedMethods: ['describe']
   Max: 30
-
 
 Style/Documentation:
   Enabled: false
