+++ lib_methods/handle_books.rb
@@ -1,0 +1,23 @@
+require './book'
+
+module HandleBook
+  def list_books
+    if @books.empty?
+      puts 'No books available'
+    else
+      @books.each do |book|
+        puts "Title: #{book.title}, Author: #{book.author}"
+      end
+    end
+  end
+
+  def create_book
+    print 'Title: '
+    title = gets.chomp
+    print 'Author: '
+    author = gets.chomp
+    book = Book.new(title, author)
+    @books.push(book)
+    puts 'Book created successfullly'
+  end
+end
